pyproject.toml
[tool.poetry]
name = "arxiv-telegram-bot"
version = "0.1.0"
description = "A Telegram bot to retrieve information on the latest research papers on Arxiv"
authors = ["Sonaal Pradeep <unofficial.sonaal@gmail.com>", "Goutham Krishna <gauthamkrishna9991@live.com>"]

[tool.poetry.dependencies]
python = "^3.8"
black = "^21.12b0"
python-telegram-bot = "^13.10"
loguru = "^0.5.3"
python-decouple = "^3.5"
arxiv = "^1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
